[runtimes] Revert the libc++ __config_site change
This is a massive revert of the following commits (from most revent to oldest):

	68fbc76a89478f901acf83fe23aa4d3506c883b7.
	529ac33197f6408952ae995075ac5e2dc5287e81
	759a6f6f737df9a830ccdac7fdc54e4c91260c10
	a386af812c52e4b26c91dcdac9e3cbbcf33f1f8a
	3e21a47fc67cf8b7674b78da2d67f97abf557663
	490a9866ce80e682624cfc8c366a9f3fca54d9dc

After checking-in the __config_site change, a lot of things started breaking
due to widespread reliance on various aspects of libc++'s build, notably the
fact that we can include the headers from the source tree, but also reliance
on various "internal" CMake variables used by the runtimes build and compiler-rt.

These were unintended consequences of the change, and after two days, we
still haven't restored all the bots to being green. Instead, now that I
understand what specific areas this will blow up in, I should be able to
chop up the patch into smaller ones that are easier to digest.

See https://reviews.llvm.org/D89041 for more details on this adventure.